% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_hierarchy.R
\name{flatten_hierarchy}
\alias{flatten_hierarchy}
\title{Flatten Hierarchy into Nodes and Edges}
\usage{
flatten_hierarchy(root_node)
}
\arguments{
\item{root_node}{The root node environment to start flattening from.}
}
\value{
A list containing `nodes` and `edges` data frames.
}
\description{
Flatten Hierarchy into Nodes and Edges
}
