% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_flow.R
\name{save_flowchart_json}
\alias{save_flowchart_json}
\title{Save flowchart data as JSON}
\usage{
save_flowchart_json(flowchart_json, out_file = "flowchart.json")
}
\arguments{
\item{flowchart_json}{A list with elements `nodes` and `edges` as produced by `parse_xlsform`.}

\item{out_file}{The file path to save the JSON to.}
}
\value{
Invisible NULL
}
\description{
Save flowchart data as JSON
}
