% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_flow.R
\name{visualize_flow}
\alias{visualize_flow}
\title{Visualize XLSForm flowchart using DiagrammeR}
\usage{
visualize_flow(flowchart_data)
}
\arguments{
\item{flowchart_json}{A list with elements `nodes` and `edges` as produced by `parse_xlsform`.}
}
\value{
A DiagrammeR graph object.
}
\description{
Visualize XLSForm flowchart using DiagrammeR
}
